finish the all quizs

+++ exercises/conversions/using_as.rs
@@ -9,7 +9,7 @@
 
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
-    total / values.len()
+    total / values.len() as f64
 }
 
 fn main() {
